Move parseHavingQuery method
