+++ frontend/src/components/JqueryQueryBuilder/utils/index.tsx
@@ -1,11 +1,5 @@
 import { JqueryQueryBuilderComps, JqueryQueryBuilderFilter } from './types';
-import { getColumnFromAlias } from '../../AdvancedHavingQueryBuilder/utils';
-import {
-  AdvancedQueryColumn,
-  AdvancedQueryFilter,
-  AdvancedQueryHaving,
-  FilterComp,
-} from '../../../types/operations';
+import { AdvancedQueryFilter, FilterComp } from '../../../types/operations';
 
 export const parseQuery = (
   finalElement: any,
@@ -30,73 +24,6 @@ export const parseQuery = (
           comp: convertJqOperatorToDDW(rulesObject[index].operator),
           value: rulesObject[index].value,
         });
-      }
-    }
-  }
-
-  return finalElement;
-};
-
-export const parseHavingQuery = (
-  finalElement: any,
-  condition: string,
-  rulesObject: any,
-  aggregateColumns?: AdvancedQueryColumn[],
-  columns?: AdvancedQueryColumn[],
-): AdvancedQueryHaving => {
-  if (rulesObject.hasOwnProperty('condition')) {
-    finalElement[`$${rulesObject.condition.toLowerCase()}`] = [];
-    finalElement = parseHavingQuery(
-      finalElement,
-      rulesObject.condition,
-      rulesObject.rules,
-      aggregateColumns,
-      columns,
-    );
-  } else {
-    for (let index = 0; index < rulesObject.length; index++) {
-      if (Array.isArray(rulesObject) && rulesObject[index].condition) {
-        finalElement[`$${condition.toLowerCase()}`].push(
-          parseHavingQuery(
-            {},
-            rulesObject[index].condition,
-            rulesObject[index].rules,
-            aggregateColumns,
-            columns,
-          ),
-        );
-      } else {
-        if (!finalElement.hasOwnProperty(`$${condition.toLowerCase()}`)) {
-          finalElement[`$${condition.toLowerCase()}`] = [];
-        }
-        if ((aggregateColumns as AdvancedQueryColumn[]).length > 0) {
-          aggregateColumns?.map((column) => {
-            if (column.name === rulesObject[index].field) {
-              finalElement[`$${condition.toLowerCase()}`].push({
-                column: rulesObject[index].field,
-                comp: convertJqOperatorToDDW(rulesObject[index].operator),
-                aggregate: column.aggregate,
-                value: { plain: rulesObject[index].value },
-              });
-            }
-          });
-        } else {
-          if (typeof rulesObject[index].value === 'number') {
-            finalElement[`$${condition.toLowerCase()}`].push({
-              column: getColumnFromAlias(rulesObject[index].field, columns as AdvancedQueryColumn[])
-                .name,
-              comp: convertJqOperatorToDDW(rulesObject[index].operator),
-              value: { plain: rulesObject[index].value },
-            });
-          } else {
-            const receivedString = rulesObject[index].value.split(',');
-            finalElement[`$${condition.toLowerCase()}`].push({
-              column: rulesObject[index].field,
-              comp: convertJqOperatorToDDW(rulesObject[index].operator),
-              value: { column: receivedString[0], aggregate: receivedString[1] },
-            });
-          }
-        }
       }
     }
   }
@@ -159,7 +86,7 @@ const convertDDWOperatorToJq = (ddwOperator: FilterComp) => {
   return operatorMap[ddwOperator];
 };
 
-const convertJqOperatorToDDW = (operator: JqueryQueryBuilderComps) => {
+export const convertJqOperatorToDDW = (operator: JqueryQueryBuilderComps): string => {
   const operatorMap = {
     equal: '$eq',
     not_equal: '$neq',
